+++ src/main/scala/apps/MatDCheck.scala
@@ -1,14 +1,14 @@
 //#!/usr/bin/env -S scala-cli shebang -Wunused:imports -Wunused:locals -deprecation
 package uni.apps
 
-//> using dep org.vastblue:uni_3:0.14.2
+//> using dep org.vastblue:uni_3:0.15.0
 
 import uni.data.*
 
 object MatDCheck {
   def main(args: Array[String]): Unit = {
     val m              = Mat.zeros[Double](3, 4) // .asInstanceOf[Mat[Double]]
-    val n: Mat[Double] = MatD(3, 4) // .asInstanceOf[Mat[Double]]
+    val n: Mat[Double] = MatD.zeros(3, 4) // .asInstanceOf[Mat[Double]]
     println(s"m: $m")
     println(s"n: $n")
     val col = m(::, 0)